requirements.txt
mysql-connector-python
numpy==1.21.2
